Yes I Know This Doesn't Work, But I Have Sleep and Other Work to do
Fixed Typo, also removed the "TODO set intake direction in OI." We need
to remove the code that sets all the motors in OI but I don't know what
code in there will or won't break things if I change it. If we just run
the commands then I think it should work, it passed the duck test for me
atleast.